% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-kanalysis.R
\name{analyze_network}
\alias{analyze_network}
\title{Network analysis using k core decomposition}
\usage{
analyze_network(namenetwork, directory = "", guild_a = "pl",
  guild_b = "pol", plot_graphs = FALSE, only_NODF = FALSE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Network analysis using k core decomposition
}
\examples{
analyze_network()
}
\keyword{cats}

