% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_graph.R
\name{matrix_graph}
\alias{matrix_graph}
\title{Plotting a matrix graph}
\usage{
matrix_graph(
  datadir,
  filename,
  style = "matrix",
  orderby = "kradius",
  label_strguilda = "Plant",
  label_strguildb = "Pollinator",
  label_size = 18,
  color_guild_a = "#4169E1",
  color_guild_b = "#FF1808",
  color_links = "grey7",
  flip_matrix = FALSE,
  links_weight = FALSE,
  show_species_names = TRUE,
  show_title = TRUE,
  show_legend = TRUE,
  print_to_file = FALSE,
  plotsdir = "plot_results/",
  plot_size = 10,
  ppi = 300,
  progress = NULL
)
}
\arguments{
\item{datadir}{the name of the file of the interaction matrix}

\item{filename}{the file with the interaction matrix}

\item{label_strguilda}{string labels of guild a}

\item{label_strguildb}{string labels of guild b}

\item{label_size}{base label size}

\item{color_guild_a}{default filling for nodes of guild_a}

\item{color_guild_b}{default filling for nodes of guild_b}

\item{color_links}{color of matrix interactions}

\item{flip_matrix}{to switch row/column guilds,}

\item{links_weight}{show links weight as a gradient of color for weighted networks}

\item{show_species_names}{add species name to number,}

\item{show_title}{show network name as plot title,}

\item{show_legend}{show guilds legend,}

\item{print_to_file}{print plot to file,}

\item{plotsdir}{default print directory}

\item{plot_size}{in inches for printed plot,}

\item{ppi}{dots per inch}

\item{progress}{for interactive visualization}

\item{orderkcoremaxby}{sets order of nodes, by kradius, kdegree or degree}
}
\description{
This function plots the matrix graph of a bipartite network. Configuration parameters and
results are stored in a global environment called mat. This environment is not destroyed
after the function is executed, so the developer can store it in a configuration file, retrieve
}
