% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-kanalysis.R
\name{read_and_analyze}
\alias{read_and_analyze}
\title{Reads and analyzes a network from a file}
\usage{
read_and_analyze(directorystr, network_file, label_strguilda, label_strguildb)
}
\arguments{
\item{directorystr}{where the network file is stored}

\item{network_file}{is the file name that contains the interaction matrix}

\item{label_strguilda}{name of the guild B of nodes}
}
\value{
\code{calc_vals} a list containing the following objects
\itemize{
 \item{\code{"result_analysis"}}{ a full network analysis}
 \item{\code{"str_guild_a"}}{ list of nodes of Guild A}
 \item{\code{"str_guild_b"}}{ list of nodes of Guild B}
 \item{\code{"name_guild_a"}}{ name of Guild A}
 \item{\code{"network_name"}}{ network name}
 }
}
\description{
Reads and analyzes a network from a file
}
\examples{
p <- read_and_analyze("../data","M_PL_003.csv","Plant","Pollinator")
}
