% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-kanalysis.R
\name{read_network}
\alias{read_network}
\title{Functions for k-core decompose a bipartite graph and computing some newtork indexes
Reads a network interaction matrix from a CSV file}
\usage{
read_network(
  namenetwork,
  guild_astr = "pl",
  guild_bstr = "pol",
  directory = ""
)
}
\description{
Args:
  namenetwork: CSV file that contains the interaction matrix
  guild_a, guild_b: Identifier of the guild of species of each class. Default "pl" (plant)
                    "pol" (pollinator)
  directory: directory where newtork CSVs are located
}
\details{
Return List:
  graph: Newtork as an igraph object
  m    : Interaction matrix
  num_guild_b : number of species of guild_b
  num_guild_a" : number of species of guild_a
  names_guild_a : names of nodes of guild_a
  names_guild_b : names of species of guild_b
}
