% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-kanalysis.R
\name{get_bipartite}
\alias{get_bipartite}
\title{Get bipartite plot}
\usage{
get_bipartite(g, str_guild_a = "Plant", str_guild_b = "Pollinator",
  plot_graphs = FALSE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Get bipartite plot
}
\examples{
get_bipartitr()
}
\keyword{cats}

