% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_graph.R
\name{polar_graph}
\alias{polar_graph}
\title{Plotting a polar plot}
\usage{
polar_graph(red, directorystr = "data/", plotsdir = "plot_results/polar/",
  print_to_file = FALSE, pshowtext = FALSE, show_histograms = TRUE,
  glabels = c("Plant", "Pollinator"), gshortened = c("pl", "pol"),
  lsize_title = 22, lsize_axis = 12, lsize_legend = 13,
  lsize_axis_title = 14, lsize_legend_title = 15, progress = NULL,
  printable_labels = 0)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Plotting a polar plot
}
\examples{
polar_graph()
}
\keyword{cats}

