% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{SVG}
\alias{SVG}
\title{SVG aux function}
\usage{
SVG(
  scale_factor,
  style = "ziggurat",
  nnodes = 50,
  flip_coordinates = FALSE,
  zoom_svgplot = 1
)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
SVG aux function
}
\examples{
SVG()
}
\keyword{cats}
