% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziggurat_graph.R
\name{ziggurat_graph}
\alias{ziggurat_graph}
\title{Plotting a ziggurat plot}
\usage{
ziggurat_graph(datadir, filename, paintlinks = TRUE,
  displaylabelszig = TRUE, print_to_file = FALSE,
  plotsdir = "plot_results/ziggurat/", flip_results = FALSE,
  aspect_ratio = 1, alpha_level = 0.2, color_guild_a = c("#4169E1",
  "#00008B"), color_guild_b = c("#F08080", "#FF0000"),
  color_link = "#696969", alpha_link = 0.2, size_link = 0.5,
  displace_y_b = rep(0, 11), displace_y_a = rep(0, 11), labels_size = 3.5,
  lsize_kcoremax = 3.5, lsize_zig = 3, lsize_kcore1 = 2.5,
  lsize_legend = 4, lsize_core_box = 2.5, labels_color = c(),
  height_box_y_expand = 1, kcore2tail_vertical_separation = 1,
  kcore1tail_disttocore = c(1, 1), innertail_vertical_separation = 1,
  horiz_kcoremax_tails_expand = 1, factor_hop_x = 1,
  displace_legend = c(0, 0), fattailjumphoriz = c(1, 1),
  fattailjumpvert = c(1, 1), coremax_triangle_height_factor = 1,
  coremax_triangle_width_factor = 1, paint_outsiders = TRUE,
  displace_outside_component = c(1, 1), outsiders_separation_expand = 1,
  outsiders_legend_expand = 1,
  weirdskcore2_horizontal_dist_rootleaf_expand = 1,
  weirdskcore2_vertical_dist_rootleaf_expand = 0,
  weirds_boxes_separation_count = 1, root_weird_expand = c(1, 1),
  hide_plot_border = TRUE, rescale_plot_area = c(1, 1),
  kcore1weirds_leafs_vertical_separation = 1, corebox_border_size = 0.2,
  kcore_species_name_display = c(), kcore_species_name_break = c(),
  shorten_species_name = 0, label_strguilda = "", label_strguildb = "",
  landscape_plot = TRUE, backg_color = "white", show_title = TRUE,
  use_spline = TRUE, spline_points = 100, svg_scale_factor = 10,
  progress = NULL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Plotting a ziggurat plot
}
\examples{
 ziggurat_graph("data/","M_PL_006.csv",print_to_file = FALSE)
}
\keyword{cats}

